Rename io.h to menuIo.h (io.cpp -> menuIo.cpp) for perevent confusing with arduino-stm32 (and other possible platforms)
