+++ src/items.h
@@ -449,7 +449,7 @@
 
   ////////////////////////////////////////////////////////////////////////
   // template implementation
-  #include "io.h"
+  #include "menuIo.h"
   #include "nav.h"
   namespace Menu {
